j'ai ajouter un route qui afficher des transactions

+++ wallet-Api-app/app/Http/Requests/StoreWalletRequest.php
@@ -23,7 +23,7 @@ class StoreWalletRequest extends FormRequest
     {
         return [
             'name'=>'required | string',
-            'balance'=>'numeric',
+            'balance'=>'integer',
             'currency'=>'required|string',
          ];
     }
